Move canStepBranch into its own function
